backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-docx
openai
passlib[bcrypt]
python-jose[cryptography]
email-validator
